+++ include/shop/Shop.hpp
@@ -40,7 +40,7 @@ public:
 	Shop(const Shop& shop) = delete;
 	Shop& operator=(const Shop& shop) = delete;
 
-	virtual std::unique_ptr<Pattern> ableSisterPattern(int pattern) const = 0;
+	virtual std::unique_ptr<Pattern> ableSisterPattern(u8 pattern) const = 0;
 };
 
 #endif
